requirements.txt
Flask==3.0.2
Pillow==10.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1 